requirements.txt
fastapi
uvicorn
yt-dlp
webvtt-py
openai-whisper
